Initiated version 0.2. Fixed constraint issue when discovering

+++ src/main/java/com/jayway/jersey/rest/resource/Resource.java
@@ -1,12 +1,15 @@
 package com.jayway.jersey.rest.resource;
 
-import com.sun.jersey.api.json.JSONJAXBContext;
-import com.sun.jersey.api.json.JSONUnmarshaller;
 import com.jayway.jersey.rest.RestfulJerseyService;
 import com.jayway.jersey.rest.constraint.Constraint;
 import com.jayway.jersey.rest.constraint.ConstraintEvaluator;
-import com.jayway.jersey.rest.dto.StringDTO;
+import com.sun.jersey.api.json.JSONJAXBContext;
+import com.sun.jersey.api.json.JSONUnmarshaller;
+import com.sun.xml.internal.bind.v2.runtime.JAXBContextImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.HttpMethod;
@@ -15,7 +18,10 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
+import javax.ws.rs.ext.MessageBodyWriter;
+import javax.ws.rs.ext.Providers;
 import javax.xml.bind.JAXBContext;
+import java.io.IOException;
 import java.io.InputStream;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
@@ -34,68 +40,84 @@ import java.util.List;
  */
 abstract public class Resource {
 
+    Logger log = LoggerFactory.getLogger( Resource.class );
 
     /**
      * lists the capabilities of this resource.
      *
      * @return string containing html
      */
-    public String capabilities( ) {
-        List<ResourceMethod> methods = new ArrayList<ResourceMethod>();
+    public void capabilitiesHtml() {
+        try {
+            ServletOutputStream os = role(HttpServletResponse.class).getOutputStream();
+            List<ResourceMethod> methods = new ArrayList<ResourceMethod>();
 
-        StringBuilder sb = new StringBuilder();
-        Class<? extends Resource> clazz = this.getClass();
-        sb.append( "<h1>" ).append( clazz.getName() ).append("</h1>");
-        for ( Method method : clazz.getDeclaredMethods() ) {
-            if ( method.isSynthetic() ) continue;
-            methods.add(new ResourceMethod(method));
-        }
-        role( HtmlHelper.class).addResourceMethods(sb, methods);
+            os.print( "<html><h1>" );
+            Class<? extends Resource> clazz = this.getClass();
+            os.print( clazz.getName() );
+            os.println( "</h1>" );
+            for ( Method method : clazz.getDeclaredMethods() ) {
+                if ( method.isSynthetic() ) continue;
+                methods.add(new ResourceMethod(method));
+            }
+            role( HtmlHelper.class).addResourceMethods( os, methods );
 
-        sb.append("<h2>Index</h2>");
-        if ( this instanceof IndexResource ) {
-            sb.append(((IndexResource) this).index());
+            os.print( "<h2>Index</h2>" );
+            if ( this instanceof IndexResource ) {
+                ResourceMethod method = findMethod("index");
+                if ( !method.isConstraintFalse() ) {
+                    try {
+                        writeJsonToStream(((IndexResource) this).index(), os);
+                    } catch ( Exception e ) {
+                        os.print( "Exception thrown when executing 'index'" );
+                    }
+                }
+            }
+            os.print("</html>");
+            os.flush();
+        } catch (IOException e) {
+            log.error("Error handling servlet output stream", e);
         }
-        return sb.toString();
     }
 
     public Response post(String method, MultivaluedMap<String,String> formParams, InputStream stream) {
         ResourceMethod m = findMethod( method );
-        if ( !m.isCommand() ) throw new WebApplicationException( HttpServletResponse.SC_NOT_FOUND );
+        if ( !m.isCommand() || m.isNotFound() || m.isConstraintFalse() || m.isIdSubResource() ) throw notFound();
 
         Object[] arguments = stream == null ? arguments(m.method, formParams ) : arguments( m.method, stream );
         return invokeCommand(m.method, this, arguments );
     }
 
 
-    private Object[] arguments( Method m, InputStream xml ) {
+    private Object[] arguments( Method m, InputStream stream ) {
         try {
             String contentType = role( HttpServletRequest.class ).getContentType();
             if ( m.getParameterTypes().length == 1 ) {
                 Class<?> dto = m.getParameterTypes()[0];
+                // TODO scrap JSON and handle simple types
                 if ( contentType.equals( MediaType.APPLICATION_JSON )) {
                     Class<?> dtoClass = dto.newInstance().getClass();
                     JAXBContext context = JSONJAXBContext.newInstance( dtoClass );
                     JSONUnmarshaller jsonUnmarshaller = JSONJAXBContext.getJSONUnmarshaller( context.createUnmarshaller() );
-                    return new Object[]{ jsonUnmarshaller.unmarshalFromJSON( xml, dtoClass ) };
+                    return new Object[]{ jsonUnmarshaller.unmarshalFromJSON( stream, dtoClass ) };
                 } else {
                     // default to xml
-                    JAXBContext context = JAXBContext.newInstance(dto.newInstance().getClass());
-                    return new Object[] { context.createUnmarshaller().unmarshal( xml ) };
+                    JAXBContext context = JAXBContextImpl.newInstance(dto.newInstance().getClass());
+                    return new Object[] { context.createUnmarshaller().unmarshal( stream ) };
                 }
             }
             return new Object[0];
         } catch (Exception e) {
-            throw new WebApplicationException( e, Response.Status.BAD_REQUEST );
+            throw badRequest( e );
         }
     }
 
     private Object[] arguments( Method m, MultivaluedMap<String, String> formParams ) {
-        if ( m.getParameterTypes().length == 1 ) {
-            Class<?> dto = m.getParameterTypes()[0];
-            return new Object[]{ populateDTO( dto, formParams, dto.getSimpleName() ) };
+        if ( m.getParameterTypes().length == 0 ) {
+            return new Object[0];
         }
-        return new Object[0];
+        Class<?> dto = m.getParameterTypes()[0];
+        return new Object[]{ populateDTO( dto, formParams, dto.getSimpleName() ) };
     }
 
     private boolean checkConstraint(Method method) {
@@ -123,28 +145,26 @@ abstract public class Resource {
     }
 
     public Resource invokePathMethod( String path ) {
-        try {
-            ResourceMethod method = findMethod( path );
-            if ( method.isSubResource() ) {
-                try {
-                    return (Resource) method.method.invoke(this);
-                } catch ( IllegalAccessException e) {
-                    e.printStackTrace();
-                    throw new WebApplicationException( Response.Status.INTERNAL_SERVER_ERROR );
-                } catch ( InvocationTargetException e) {
-                    if ( e.getCause() instanceof RuntimeException ) {
-                        throw (RuntimeException) e.getCause();
-                    }
-                    e.printStackTrace();
-                    throw new WebApplicationException( Response.Status.INTERNAL_SERVER_ERROR );
+        ResourceMethod method = findMethod( path );
+        if ( method.isSubResource() ) {
+            try {
+                return (Resource) method.method.invoke(this);
+            } catch ( IllegalAccessException e) {
+                e.printStackTrace();
+                throw internalServerError( e );
+            } catch ( InvocationTargetException e) {
+                if ( e.getCause() instanceof RuntimeException ) {
+                    throw (RuntimeException) e.getCause();
                 }
+                log.error("Error invoking resource method", e);
+                throw internalServerError( e );
             }
-        } catch ( WebApplicationException notFound ) {
-            if ( this instanceof IdResource ) {
-                return ((IdResource) this).id( new StringDTO( path ) );
-            }
+        } else if ( ( method.isIdSubResource() || method.isNotFound() ) && this instanceof IdResource ) {
+            ResourceMethod id = findMethod("id");
+            if ( id.isConstraintFalse() ) throw notFound();
+            return ((IdResource) this).id( path );
         }
-        throw new WebApplicationException( Response.Status.NOT_FOUND );
+        throw notFound();
     }
 
 
@@ -161,19 +181,21 @@ abstract public class Resource {
     public ResourceMethod findMethod( String name ) {
         Class<? extends Resource> clazz = this.getClass();
         for ( Method method : clazz.getDeclaredMethods() ) {
+            if ( method.isSynthetic() ) continue;
             if ( method.getName().equals( name ) ) {
-                ResourceMethod resourceMethod = new ResourceMethod( method );
-                if ( !resourceMethod.isIllegal() ) return resourceMethod;
+                return new ResourceMethod( method );
             }
         }
-        throw new WebApplicationException( Response.Status.NOT_FOUND );
+        return new ResourceMethod();
     }
 
     public void invokeDelete() {
         if ( this instanceof DeletableResource) {
+            ResourceMethod delete = findMethod("delete");
+            if ( delete.isConstraintFalse() ) throw notFound();
             (( DeletableResource) this).delete();
         } else {
-            throw new WebApplicationException( HttpServletResponse.SC_METHOD_NOT_ALLOWED );
+            throw notAllowed();
         }
     }
 
@@ -183,10 +205,10 @@ abstract public class Resource {
             return successResponse();
         } catch (InvocationTargetException e) {
             if ( e.getCause() instanceof WebApplicationException ) throw (WebApplicationException) e.getCause();
-            throw new WebApplicationException( e.getCause(), Response.Status.BAD_REQUEST );
+            throw badRequest( e );
         } catch (IllegalAccessException e) {
-            e.printStackTrace();
-            throw new WebApplicationException( e.getCause(), Response.Status.INTERNAL_SERVER_ERROR );
+            log.error("Could not access command", e);
+            throw internalServerError( e );
         }
     }
 
@@ -194,11 +216,27 @@ abstract public class Resource {
         return Response.ok( "Operation Completed Successfully" ).status(200).build();
     }
 
+    /**
+     * Special case for handling html. Data object are converted
+     * to JSON when serializing to HTML.
+     *
+     * @param get
+     * @throws IOException
+     */
+    public void getHtml( String get ) throws IOException {
+        writeJsonToStream( get(get), role(HttpServletResponse.class).getOutputStream());
+    }
+
+    private void writeJsonToStream(Object result, ServletOutputStream os) throws IOException {
+        Providers ps = role(Providers.class);
+        MessageBodyWriter uw = ps.getMessageBodyWriter(result.getClass(), result.getClass(), new Annotation[0], MediaType.APPLICATION_JSON_TYPE);
+        uw.writeTo(result, result.getClass(), result.getClass(), new Annotation[0], MediaType.APPLICATION_JSON_TYPE, null, os );
+        os.flush();
+    }
+
     public Object get( String get ) {
         ResourceMethod m = findMethod(get);
-        if ( m.isSubResource() ) {
-            throw new WebApplicationException( Response.Status.NOT_FOUND );
-        }
+        if ( m.isSubResource() || m.isNotFound() || m.isConstraintFalse() || m.isIdSubResource() ) throw notFound();
 
         if ( m.isCommand() ) {
             Response response = Response.status( HttpServletResponse.SC_METHOD_NOT_ALLOWED).entity( new HtmlHelper().createForm( m.method, HttpMethod.POST ) ).build();
@@ -210,21 +248,16 @@ abstract public class Resource {
             return new HtmlHelper().createForm(m.method, HttpMethod.GET );
         } else {
             try {
-                Object result;
-                if ( m.method.getParameterTypes().length == 1) {
-                    Class<?> dto = m.method.getParameterTypes()[0];
-                    result = m.method.invoke(this, populateDTO( dto, queryParams, dto.getSimpleName() ) );
-                } else {
-                    result = m.method.invoke( this );
+                if ( m.method.getParameterTypes().length == 0 ) {
+                    return m.method.invoke( this );
                 }
-                if ( role(HttpServletRequest.class).getHeader("Accept").contains(MediaType.TEXT_HTML) ) {
-                    return result.toString();
-                }
-                return result;
+                // TODO handle simple types
+                Class<?> dto = m.method.getParameterTypes()[0];
+                return m.method.invoke(this, populateDTO( dto, queryParams, dto.getSimpleName() ) );
             } catch ( IllegalAccessException e) {
-                throw new WebApplicationException( e, Response.Status.INTERNAL_SERVER_ERROR );
+                throw internalServerError( e );
             } catch ( InvocationTargetException e) {
-                throw new WebApplicationException( e, Response.Status.INTERNAL_SERVER_ERROR );
+                throw internalServerError( e );
             }
         }
     }
@@ -255,24 +288,42 @@ abstract public class Resource {
             }
             return o;
         } catch (Exception e) {
-            throw new WebApplicationException( e, Response.Status.BAD_REQUEST );
+            throw badRequest(e);
         }
     }
 
+    private WebApplicationException notFound() {
+        return new WebApplicationException( Response.Status.NOT_FOUND );
+    }
+
+    private WebApplicationException notAllowed() {
+        return new WebApplicationException( HttpServletResponse.SC_METHOD_NOT_ALLOWED );
+    }
+
+    private WebApplicationException internalServerError( Exception e ) {
+        return new WebApplicationException( e, Response.Status.INTERNAL_SERVER_ERROR );
+    }
+
+    private WebApplicationException badRequest( Exception e ) {
+        return new WebApplicationException( e, Response.Status.BAD_REQUEST );
+    }
 
     class ResourceMethod {
-        private MethodType type = MethodType.ILLEGAL;
+        private MethodType type = MethodType.NOT_FOUND;
         private Method method;
         private String name;
-        private boolean isIndex;
 
+        public ResourceMethod() {}
         public ResourceMethod( Method method ) {
             this.method = method;
             this.name = method.getName();
 
             if ( Modifier.isAbstract( method.getModifiers()) ) return;
             if ( !Modifier.isPublic( method.getModifiers())) return;
-            if ( !checkConstraint(method) ) return;
+            if ( !checkConstraint(method) ) {
+                type = MethodType.CONSTRAINT_FALSE;
+                return;
+            }
 
             if ( argumentCheck( method.getParameterTypes() ) ) {
                 handleReturnType( method.getReturnType() );
@@ -288,12 +339,9 @@ abstract public class Resource {
                 type = MethodType.COMMAND;
             } else if ( Resource.class.isAssignableFrom( returnType ) ) {
                 if ( method.getParameterTypes().length == 0 ) type = MethodType.SUBRESOURCE;
-                else type = MethodType.ILLEGAL;
+                else type = MethodType.ID_RESOURCE;
             } else {
                 type = MethodType.QUERY;
-                if ( name.equals( "index" ) ){
-                    isIndex = true;
-                }
             }
         }
         public MethodType type() {
@@ -312,16 +360,19 @@ abstract public class Resource {
         public boolean isSubResource() {
             return type == MethodType.SUBRESOURCE;
         }
-        public boolean isIllegal() {
-            return type == MethodType.ILLEGAL;
+        public boolean isNotFound() {
+            return type == MethodType.NOT_FOUND;
         }
-        public boolean isIndex() {
-            return isIndex;
+        public boolean isConstraintFalse() {
+            return type == MethodType.CONSTRAINT_FALSE;
+        }
+        public boolean isIdSubResource() {
+            return type == MethodType.ID_RESOURCE;
         }
     }
 
     private enum MethodType {
-        COMMAND, QUERY, SUBRESOURCE, ILLEGAL
+        COMMAND, QUERY, SUBRESOURCE, NOT_FOUND, CONSTRAINT_FALSE, ID_RESOURCE
     }
 
 }
